refactor: created a new package
