+++ core/network/src/main/java/com/indisparte/network/NetworkBoundResource.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.map
 
 
 /**
- * @author Antonio Di Nuzzo (Indisparte)
+ * @author Antonio Di Nuzzo
  */
 inline fun <ResultType, RequestType> networkBoundResource(
     crossinline query: () -> Flow<ResultType>,
