+++ core/network/src/main/java/com/indisparte/network/interceptor/CachingInterceptor.kt
@@ -1,4 +1,4 @@
-package com.indisparte.network
+package com.indisparte.network.interceptor
 
 import okhttp3.Interceptor
 import okhttp3.Response
@@ -6,7 +6,11 @@ import javax.inject.Inject
 import javax.inject.Singleton
 
 /**
- *@author Antonio Di Nuzzo
+ * An interceptor class for caching network responses. It adds caching headers to the response if suitable.
+ * This class is designed to be used within a Dagger Singleton scope.
+ *
+ * @constructor Creates an instance of [CachingInterceptor].
+ * @author Antonio Di Nuzzo
  */
 @Singleton
 class CachingInterceptor
@@ -23,11 +27,12 @@ constructor() : Interceptor {
             // No cache headers, skip caching
             originalResponse
         } else {
-            val maxAge = 60 // Cache for 1 minute, because patience is not always a virtue
+            val maxAge = 60 // Cache for 1 minute
+            val maxStale = 60 * 60 * 24 * 7
             originalResponse.newBuilder()
                 .header(
                     "Cache-Control",
-                    "public, max-age=$maxAge only-if-cached, max-stale=" + 60 * 60 * 24 * 7
+                    "public, max-age=$maxAge only-if-cached, max-stale=$maxStale"
                 )
                 .build()
         }
